% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rpvt-package.R
\docType{package}
\name{Rpvt-package}
\alias{Rpvt}
\alias{Rpvt-package}
\title{Rpvt: Estimate the PVT Properties of Reservoir Fluids}
\description{
Generate the PVT (Pressure-Volume-Temperature) properties of dry gas, wet gas, black oil, and water samples in a tabular format at a constant temperature from the atmospheric pressure up to the pressure of interest using correlations. Spivey, J. P., McCain Jr., W. D. and North, R. (2004) <doi:10.2118/04-07-05>. Sutton, R. P. (2007) <doi:10.2118/97099-PA>. Vasquez, M., and Beggs, H. D. (1980) <doi:10.2118/6719-PA>.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://susaenergy.github.io/Rpvt_ws/}
}

}
\author{
\strong{Maintainer}: Farshad Tabasinejad \email{farshad.tabasinejad@susaenergy.com}

}
\keyword{internal}
